refactor(graph-db-server): remove deep relative imports
