+++ packages/systems/graph-db-server/src/application/workflows/sessionState.ts
@@ -10,7 +10,7 @@ import { getGraph } from '@vt/graph-db-server/state/graph-store'
 import { getProjectRootWatchedDirectory } from '@vt/graph-db-server/state/watch-folder-store'
 import { getReadPaths, getVaultPaths, getWritePath } from '@vt/graph-db-server/state/vaultAllowlist'
 import { getFolderStateForActiveView } from '@vt/graph-db-server/views/folderStateOps'
-import { getFolderTreeReadModel } from '../../state/folder-tree-read-model-store.ts'
+import { getFolderTreeReadModel } from '@vt/graph-db-server/state/folder-tree-read-model-store'
 import { handleReadSessionState } from '../core/handleSessionState.ts'
 import { jsonResult, notFoundResult, type HttpResult } from './httpResult.ts'
 import type { WorkflowSessionRegistry } from './sessionRoutes.ts'
