+++ packages/systems/graph-db-server/src/daemon/index.ts
@@ -1,4 +1,4 @@
-export * from './contract.ts'
+export * from '../contract.ts'
 export * from './portFile.ts'
 export {
   claimDaemonOwner,
