+++ packages/systems/graph-db-server/src/daemon/startDaemon.ts
@@ -2,7 +2,7 @@ import { mkdir } from 'node:fs/promises'
 import { join, resolve } from 'node:path'
 import { SpanStatusCode, trace, type Span } from '@opentelemetry/api'
 import { SessionRegistry } from '../application/session/registry.ts'
-import { CONTRACT_VERSION } from './contract.ts'
+import { CONTRACT_VERSION } from '../contract.ts'
 import { createDaemonApp } from '../routes/daemonApp.ts'
 import type { BoundDaemonHttpServer } from './daemonHttpServer.ts'
 import type { DaemonWatcherController } from './lifecycle/daemonWatcherLifecycle.ts'
