+++ packages/systems/graph-db-server/src/routes/__tests__/folder/folderState.test.ts
@@ -2,7 +2,7 @@ import { afterEach, beforeEach, describe, expect, test } from 'vitest'
 import { mkdtemp, rm } from 'node:fs/promises'
 import { tmpdir } from 'node:os'
 import { join } from 'node:path'
-import { SessionCreateResponseSchema } from '../../../daemon/contract.ts'
+import { SessionCreateResponseSchema } from '@vt/graph-db-server/contract'
 import { type DaemonHandle, startDaemon } from '../../../daemon/server.ts'
 import { SessionRegistry } from '../../../application/session/registry.ts'
 import { createDaemonApp } from '../../daemonApp.ts'
