+++ packages/systems/graph-db-server/src/routes/__tests__/folder/sseNoFolderTreeRescan.test.ts
@@ -15,7 +15,7 @@ import {
   type GraphNode,
 } from '@vt/graph-model/graph'
 import { type DaemonHandle, startDaemon } from '../../../daemon/server.ts'
-import { SessionCreateResponseSchema } from '../../../daemon/contract.ts'
+import { SessionCreateResponseSchema } from '@vt/graph-db-server/contract'
 import type { ProjectedGraph } from '@vt/graph-state/contract'
 import { getGraph, setGraph } from '../../../state/graph-store.ts'
 import { publish } from '../../../state/events/deltaEventBus.ts'
