+++ packages/systems/graph-db-server/src/routes/__tests__/folder/sseSubscriptionClient.test.ts
@@ -3,7 +3,7 @@ import { mkdtemp, rm, writeFile } from 'node:fs/promises'
 import { tmpdir } from 'node:os'
 import { join } from 'node:path'
 import { type DaemonHandle, startDaemon } from '../../../daemon/server.ts'
-import { SessionCreateResponseSchema } from '../../../daemon/contract.ts'
+import { SessionCreateResponseSchema } from '@vt/graph-db-server/contract'
 import type { ProjectedGraph } from '@vt/graph-state/contract'
 
 async function withTempVault(): Promise<string> {
