+++ packages/systems/graph-db-server/src/routes/__tests__/session/sessionEvents.test.ts
@@ -12,7 +12,7 @@ import {
 } from '@vt/graph-model/graph'
 import { type DaemonHandle, startDaemon } from '../../../daemon/server.ts'
 import { SessionRegistry } from '../../../application/session/registry.ts'
-import { SessionCreateResponseSchema } from '../../../daemon/contract.ts'
+import { SessionCreateResponseSchema } from '@vt/graph-db-server/contract'
 import type { ProjectedGraph } from '@vt/graph-state/contract'
 import { getGraph, setGraph } from '../../../state/graph-store.ts'
 import { publish } from '../../../state/events/deltaEventBus.ts'
