+++ packages/systems/graph-db-server/tests/sse-replay-buffer.test.ts
@@ -10,7 +10,7 @@ import {
   type GraphNode,
 } from '@vt/graph-model/graph'
 import type { ProjectedGraph } from '@vt/graph-state/contract'
-import { SessionCreateResponseSchema } from '../src/daemon/contract.ts'
+import { SessionCreateResponseSchema } from '@vt/graph-db-server/contract'
 import { publish } from '../src/state/events/deltaEventBus.ts'
 import { type DaemonHandle, startDaemon } from '../src/daemon/server.ts'
 import { getGraph, setGraph } from '../src/state/graph-store.ts'
