+++ packages/systems/graph-db-server/tests/write-markdown-file-roundtrip.test.ts
@@ -5,7 +5,7 @@ import { afterEach, beforeEach, describe, expect, test } from 'vitest'
 import { createEmptyGraph, type GraphNode } from '@vt/graph-model'
 import type { ProjectedGraph } from '@vt/graph-state/contract'
 
-import { SessionCreateResponseSchema } from '../src/daemon/contract.ts'
+import { SessionCreateResponseSchema } from '@vt/graph-db-server/contract'
 import { startDaemon, type DaemonHandle } from '../src/daemon/index.ts'
 import { setGraph } from '../src/state/graph-store.ts'
 import { clearWatchFolderState } from '../src/state/watch-folder-store.ts'
